+++ src/dropdown.js
@@ -10,8 +10,9 @@ const DEFAULT_CLASS_NAME = 'dropdown-menu textcomplete-dropdown';
 
 /**
  * @typedef {Object} Dropdown~Offset
- * @prop {number} top
- * @prop {number} left
+ * @prop {number} [top]
+ * @prop {number} [left]
+ * @prop {number} [right]
  */
 
 /**
@@ -30,7 +31,6 @@ class Dropdown extends EventEmitter {
     el.id = uniqueId('textcomplete-dropdown-');
     extend(el.style, {
       display: 'none',
-      left: 0,
       position: 'absolute',
       zIndex: 10000,
     });
@@ -176,8 +176,11 @@ class Dropdown extends EventEmitter {
    * @returns {this}
    */
   setOffset(cursorOffset) {
-    this.el.style.top = `${cursorOffset.top}px`;
-    this.el.style.left = `${cursorOffset.left}px`;
+    ['top', 'right', 'bottom', 'left'].forEach(name => {
+      if (cursorOffset.hasOwnProperty(name)) {
+        this.el.style[name] = `${cursorOffset[name]}px`;
+      }
+    });
     return this;
   }