+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This change log adheres to [keepachangelog.com](http://keepachangelog.com).
 ## [Unreleased]
 ### Added
 - Enable to select dropdown in touch devices.
+- Enable to use markdown in jsdoc.
 
 ## [0.1.2] - 2016-02-22
 ### Added